requirements.txt
# GLOCAL-FSL-OOD Requirements

# Core
torch>=2.0.0
torchvision>=0.15.0

# CLIP
open_clip_torch>=2.20.0

# LLM / Transformers
transformers>=4.36.0
accelerate>=0.25.0
bitsandbytes>=0.41.0  # Optional: for quantized LLM loading

# Data
numpy>=1.24.0
pandas>=2.0.0
Pillow>=9.5.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Logging & Visualization
pyyaml>=6.0
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0
tensorboard>=2.14.0
# wandb>=0.16.0  # Optional

# Utilities
requests>=2.31.0
einops>=0.7.0
timm>=0.9.0  # Optional: additional vision models
